requirements.txt
autogen-agentchat
streamlit
python-dotenv
google-generativeai
pyautogen
ag2[gemini]
faster-whisper